Initial commit.
I was editing a file, when all of a sudden, my system complained that
the disk was full. So the file didn't write correctly, and I lost the
file. It was at this moment that I realized: "Holy guacamole! I need to
put this in a git repository!". So this is what I'm now doing.

PS: Vim saved my life by caching the file.